Remove global variables that are difficult to follow through the code
Before the code was wrapped in a big user.then, but it make it difficult
to follow. Promises is good for creating complex but easy to read flow
control. You can see all the small chunk of code and see what they depend
on.

If the depend on multiple thing, the function Promsise.all([]) can be used