getAppBillableHours should return the correct amount when giving a timespan
Before there was sereral cases, where giving a timespan would not return
the correct hours.

This takes into account that you can't show hours that are into the future.

If a end timespan is in the future, it should only return hours that are
from now()

This simplifies it since it will only return once.